+++ soccer-fyrox/src/game_global.rs
@@ -191,20 +191,20 @@ impl GameGlobal {
                 };
 
                 self.media
-                    .blit_image(scene, "menu", &[image_i1, image_i2], 0, 0, DRAW_MENU_Z);
+                    .blit_image(scene, "menu", &[image_i1, image_i2], 0., 0., DRAW_MENU_Z);
             }
             Play => {
                 //# Display score bar at top
                 self.media
-                    .blit_image(scene, "bar", &[], HALF_WINDOW_W - 176, 0, DRAW_GAME_HUD_Z);
+                    .blit_image(scene, "bar", &[], HALF_WINDOW_W - 176., 0., DRAW_GAME_HUD_Z);
                 //# Show score for each team
                 for i in 0..2 {
                     self.media.blit_image(
                         scene,
                         "s",
                         &[self.game.teams[i].score],
-                        HALF_WINDOW_W + 7 - 39 * (i as i16),
-                        6,
+                        HALF_WINDOW_W + 7. - 39. * (i as f32),
+                        6.,
                         DRAW_GAME_SCORES_Z,
                     );
                 }
@@ -215,8 +215,8 @@ impl GameGlobal {
                         scene,
                         "goal",
                         &[],
-                        HALF_WINDOW_W - 300,
-                        HEIGHT / 2 - 88,
+                        HALF_WINDOW_W - 300.,
+                        HEIGHT / 2. - 88.,
                         DRAW_GAME_HUD_Z,
                     );
                 }
@@ -225,7 +225,7 @@ impl GameGlobal {
                 //# Display "Game Over" image
                 let index = (self.game.teams[1].score > self.game.teams[0].score) as u8;
                 self.media
-                    .blit_image(scene, "over", &[index], 0, 0, DRAW_GAME_OVER_Z);
+                    .blit_image(scene, "over", &[index], 0., 0., DRAW_GAME_OVER_Z);
 
                 //# Show score for each team
                 for i in 0..2 {
@@ -233,8 +233,8 @@ impl GameGlobal {
                         scene,
                         "l",
                         &[i as u8, self.game.teams[i as usize].score],
-                        HALF_WINDOW_W + 25 - 125 * i,
-                        144,
+                        HALF_WINDOW_W + 25. - 125. * i as f32,
+                        144.,
                         DRAW_GAME_OVER_Z,
                     );
                 }
@@ -249,7 +249,7 @@ impl GameGlobal {
             .with_projection(Projection::Orthographic(OrthographicProjection {
                 z_near: CAMERA_NEAR_Z,
                 z_far: CAMERA_FAR_Z,
-                vertical_size: (HEIGHT / 2) as f32,
+                vertical_size: (HEIGHT / 2.),
             }))
             .build(&mut scene.graph)
     }
